+++ packages/yourloops/components/patient-list/custom-cells.tsx
@@ -37,13 +37,13 @@ import AccessTimeIcon from '@mui/icons-material/AccessTime'
 import FlagIcon from '@mui/icons-material/Flag'
 import FlagOutlineIcon from '@mui/icons-material/FlagOutlined'
 import { type Patient } from '../../lib/patient/models/patient.model'
-import { type MonitoringAlerts } from '../../lib/patient/models/monitoring-alerts.model'
 import Badge from '@mui/material/Badge'
 import { TimeSpentOufOfRangeIcon } from '../icons/diabeloop/time-spent-ouf-of-range-icon'
 import { NoDataIcon } from '../icons/diabeloop/no-data-icon'
 import { HypoglycemiaIcon } from '../icons/diabeloop/hypoglycemia-icon'
 import { NoMessageIcon } from '../icons/diabeloop/no-message-icon'
 import { MessageIcon } from '../icons/diabeloop/message-icon'
+import { Unit } from 'medical-domain'
 
 interface FlagCellProps {
   isFlagged: boolean
@@ -51,7 +51,7 @@ interface FlagCellProps {
 }
 
 interface MonitoringAlertsCellProps {
-  monitoringAlerts: MonitoringAlerts
+  patient: Patient
 }
 
 interface MessageCellProps {
@@ -105,30 +105,86 @@ export const PendingIconCell: FunctionComponent = () => {
   )
 }
 
-export const MonitoringAlertsCell: FunctionComponent<MonitoringAlertsCellProps> = ({ monitoringAlerts }) => {
+export const MonitoringAlertsCell: FunctionComponent<MonitoringAlertsCellProps> = ({ patient }) => {
   const { t } = useTranslation()
   const theme = useTheme()
+  const monitoringAlerts = patient.monitoringAlerts
 
+  // This default value assignment will be removed in YLP-2324
+  // (it is only necessary because monitoring features have not yet been enabled for every patient)
+  const monitoringParameters = patient.monitoring?.parameters ?? {
+    bgUnit: Unit.MilligramPerDeciliter,
+    lowBg: 70,
+    highBg: 100,
+    outOfRangeThreshold: 10,
+    veryLowBg: 65,
+    hypoThreshold: 5,
+    nonDataTxThreshold: 15,
+    reportingPeriod: 55
+  }
+  const unit = monitoringParameters.bgUnit
   const isTimeSpentAwayFromTargetAlertActive = monitoringAlerts.timeSpentAwayFromTargetActive
   const isFrequencyOfSevereHypoglycemiaAlertActive = monitoringAlerts.frequencyOfSevereHypoglycemiaActive
   const isNonDataTransmissionAlertActive = monitoringAlerts.nonDataTransmissionActive
+  const timeSpentAwayFromTargetRate = Math.round(monitoringAlerts.timeSpentAwayFromTargetRate * 10) / 10
+  const frequencyOfSevereHypoglycemiaRate = Math.round(monitoringAlerts.frequencyOfSevereHypoglycemiaRate * 10) / 10
+  const nonDataTransmissionRate = Math.round(monitoringAlerts.nonDataTransmissionRate * 10) / 10
+  const sharedTooltip = t('monitoring-alerts-shared-tooltip')
 
   return (
     <>
-      <Tooltip title={t('time-out-of-range-target-tooltip')} data-testid="time-spent-out-of-range-icon-tooltip">
-          <TimeSpentOufOfRangeIcon
-            color={isTimeSpentAwayFromTargetAlertActive ? 'inherit' : 'disabled'}
-            data-testid="time-spent-out-of-range-icon"
-          />
+      <Tooltip
+        title={
+          <>
+            <Box>{t('time-out-of-range-target-tooltip1', { percentage: timeSpentAwayFromTargetRate })}</Box>
+            <Box>
+              {t('time-out-of-range-target-tooltip2', {
+                min: monitoringParameters.lowBg,
+                max: monitoringParameters.highBg,
+                threshold: monitoringParameters.outOfRangeThreshold,
+                unit
+              })}
+            </Box>
+            <Box>{sharedTooltip}</Box>
+          </>
+        }
+        data-testid="time-spent-out-of-range-icon-tooltip"
+      >
+        <TimeSpentOufOfRangeIcon
+          color={isTimeSpentAwayFromTargetAlertActive ? 'inherit' : 'disabled'}
+          data-testid="time-spent-out-of-range-icon"
+        />
       </Tooltip>
-      <Tooltip title={t('hypoglycemia-tooltip')}>
+      <Tooltip
+        title={
+          <>
+            <Box>{t('hypoglycemia-tooltip1', { percentage: frequencyOfSevereHypoglycemiaRate })}</Box>
+            <Box>
+              {t('hypoglycemia-tooltip2', {
+                veryLowBg: monitoringParameters.veryLowBg,
+                threshold: monitoringParameters.hypoThreshold,
+                unit
+              })}
+            </Box>
+            <Box>{sharedTooltip}</Box>
+          </>
+        }
+      >
         <HypoglycemiaIcon
           sx={{ marginLeft: theme.spacing(1) }}
           color={isFrequencyOfSevereHypoglycemiaAlertActive ? 'error' : 'disabled'}
           data-testid="hypoglycemia-icon"
         />
       </Tooltip>
-      <Tooltip title={t('data-not-transferred-tooltip')}>
+      <Tooltip
+        title={
+          <>
+            <Box>{t('data-not-transferred-tooltip1', { percentage: nonDataTransmissionRate })}</Box>
+            <Box>{t('data-not-transferred-tooltip2', { threshold: monitoringParameters.nonDataTxThreshold })}</Box>
+            <Box>{sharedTooltip}</Box>
+          </>
+        }
+      >
         <NoDataIcon
           sx={{ marginLeft: theme.spacing(1) }}
           color={isNonDataTransmissionAlertActive ? 'inherit' : 'disabled'}